backend/requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0

# Data Validation
pydantic>=2.9.0
pydantic-settings>=2.6.0

# ML/AI
torch>=2.5.0
numpy>=2.1.0

# Data Processing
pandas>=2.2.0
openpyxl>=3.1.5
xlrd>=2.0.1
python-magic>=0.4.27
chardet>=5.2.0

# Utilities
python-multipart>=0.0.17
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Development
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
black>=24.10.0
flake8>=7.1.0
